requirements.txt
streamlit==1.41.1
pandas==2.2.3
plotly==5.24.1
mysql-connector-python==9.1.0
python-dotenv==1.0.1
sqlalchemy==2.0.35
